+++ SadnaSrc/SadnaSrc/StoreCenter/StoreShoppingService.cs
@@ -58,14 +58,13 @@ namespace SadnaSrc.StoreCenter
 		    return slave.answer;
 	    }
 
-	    public MarketAnswer SearchProduct(string type, string value, int minPrice, int maxPrice, string category)
+	    public MarketAnswer SearchProduct(string type, string value, double minPrice, double maxPrice, string category)
 	    {
 		    SearchProductSlave slave = new SearchProductSlave(_shopper, storeDL);
 		    slave.SearchProduct(type, value,minPrice, maxPrice, category);
 		    return slave.Answer;
 
 	    }
-	   
 
 	    public MarketAnswer AddProductToCart(string store, string productName, int quantity)
         {
